+++ src/RestApi/Applet/Version/AppletVersionManagement.ts
@@ -7,6 +7,7 @@ import IAppletVersionFilter, { IAppletVersionListFilter } from './IAppletVersion
 import AppletVersionFileManagement from './File/AppletVersionFileManagement';
 import { checkAppletVersionUpdatable, checkAppletVersionCreatable } from './AppletVersionValidator';
 import { Dependencies, createDependencies } from '../../Dependencies';
+import { PaginatedList } from '../../../Lib/Pagination/PaginatedList';
 
 export const RESOURCE: string = 'version';
 
@@ -19,11 +20,9 @@ export default class AppletVersionManagement {
 		this.file = new AppletVersionFileManagement(this.dependencies);
 	}
 
-	public async list(appletUid: string, filter: IAppletVersionListFilter = {}): Promise<IAppletVersion[]> {
+	public async list(appletUid: string, filter: IAppletVersionListFilter = {}): Promise<PaginatedList<IAppletVersion>> {
 		const response = await getResource(this.options, AppletVersionManagement.getResource(appletUid), filter);
-		const data: IAppletVersion[] = await parseJSONResponse(response);
-
-		return data.map((item: IAppletVersion) => new AppletVersion(item));
+		return this.dependencies.paginator.getPaginatedListFromResponse(response, (item: IAppletVersion) => new AppletVersion(item));
 	}
 
 	public async get(appletUid: string, version: string, filter: IAppletVersionFilter = {}): Promise<IAppletVersion> {